gen_release accepts OS name

+++ gen_release.sh
@@ -1,5 +1,12 @@
 #!/bin/bash
 
+if [ "$#" -ne 1 ]; then
+    echo "Usage: ./gen_release.sh OS_NAME"
+    exit 1
+fi
+
+OS=$1
+
 RELEASE=release
 if [ -d "$RELEASE" ]; then
     rm -rf $RELEASE
@@ -9,7 +16,7 @@ mkdir $RELEASE
 
 cd $RELEASE
 cmake ..
-make
+make VERBOSE=1
 mkdir kallisto
 
 cp src/kallisto kallisto
@@ -18,5 +25,7 @@ cp -r ../test kallisto
 cp ../license.txt kallisto
 cp ../README.md kallisto
 
-tar -cvvf kallisto.tar kallisto
-gzip kallisto.tar
+TAR_NAME="kallisto_${OS}.tar"
+
+tar -cvvf $TAR_NAME kallisto
+gzip $TAR_NAME
